+++ test/import-test-explicit-error-no-export/playwright.config.ts
@@ -1,0 +1,13 @@
+import { defineConfig } from '@playwright/test';
+import { defineBddConfig } from 'playwright-bdd';
+
+const testDir = defineBddConfig({
+  features: 'features',
+  steps: 'steps/steps.ts',
+  importTestFrom: 'steps/fixtures.ts',
+  disableWarnings: { importTestFrom: true },
+});
+
+export default defineConfig({
+  testDir,
+});