+++ test/import-test-explicit-error-no-export/steps/fixtures.ts
@@ -1,0 +1,8 @@
+import { test as base, createBdd } from 'playwright-bdd';
+
+// missing export
+const test = base.extend<{ option1: string }>({
+  option1: ['foo', { option: true }],
+});
+
+export const { Given } = createBdd(test);