+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## dev
+* fix: check exported test variable for manual importTestFrom, ([#196](https://github.com/vitalets/playwright-bdd/issues/196))
+
 ## 7.1.1
 * fix: improve error messages when guessing importTestFrom, ([#196](https://github.com/vitalets/playwright-bdd/issues/196))
 * fix: bddgen env throws if there is no @cucumber/cucumber